XMPPJSONPatch implementation
This OpenUI5 data model uses a node on an XMPP server to synchronize itself to different connected clients.
Because of the usage of CRDT (Conflict-free replicated data type) it is easy to build offline capabilities on top of it.

The XMPP server requires the following:

 * PubSub Support (XEP-0060)
 * WebSocket (RFC 7395)
 * Anonymous access and node creation capabilities

It was developed using ejabbered with the following configuration changes:

/etc/ejabberd/ejabberd.yml
...
anonymous_protocol: both
...
host_config:
   "localhost":
     auth_method:
       - internal
       - anonymous

The wireformat used in JSONPatch.